Module Asai.Explicator

Turning location information into highlighted text suitable for rendering. You probably do not need this module unless you want to create your own diagnostic handler.

exception Invalid_range of Range.t * [ `Begin of [ `Offset of [ `Negative of int | `Beyond_end_of_file of int * int | `Within_newline of int * (int * int) ] | `Incorrect_start_of_line of int * int | `Incorrect_line_num of int * int ] | `End of [ `Offset of [ `Negative of int | `Beyond_end_of_file of int * int | `Within_newline of int * (int * int) ] | `Incorrect_start_of_line of int * int | `Incorrect_line_num of int * int ] | `End_of_file of [ `Offset of [ `Negative of int | `Beyond_end_of_file of int * int | `Within_newline of int * (int * int) ] | `Incorrect_start_of_line of int * int | `Incorrect_line_num of int * int ] | `Not_end_of_file of int * int ]

Invalid_range (range, reason) means that range is an invalid range because of reason. This exception will be raised only when the debug mode is enabled. See the debug argument of Explicator.S.explicate.

module type Tag = sig ... end

The signature of tags

module type S = sig ... end

The signature of explicators.

val default_blend : priority:('tag -> int) -> 'tag -> 'tag -> 'tag

The default tag blending algorithm that chooses the more important tag based on priority.

module Make (Tag : Tag) : S with module Tag := Tag

Making an explicator.