Module type Reporter.Message

The signature of messages. An implementer should specify the message type used in their library or application.

type t

The type of all messages.

val default_severity : t -> Diagnostic.severity

The default severity level of a message. Severity levels classify diagnostics into errors, warnings, etc. It is about how serious the end user should take the diagnostic, not whether the program should stop or continue. The severity may be overwritten at the time of issuing a diagnostic.

val short_code : t -> string

A concise, ideally Google-able string representation of each message. Detailed or long descriptions should be avoided---the shorter, the better. For example, E001 works better than type-checking error. It will be assumed that the string representation has no control characters (such as newline characters).