Module Asai.SourceReader

Reading the source content. It uses memory-mapped I/O for files. You probably do not need this module unless you want to create your own diagnostic handler.

type source

An abstract type of sources.

val load : Range.source -> source

load source loads the source.

val length : source -> int

length source gets the size of the source content.

val unsafe_get : source -> int -> char

unsafe_get source i reads the ith byte of the file without checking the file size.

val run : (unit -> 'a) -> 'a

run f runs the thunk f and handles the internal algebraic effects.