Module Asai.Tty

Diagnostic display for UNIX terminals.

Display

alert unstable The TTY handler will likely change significantly in the future to account for more features.

module Make (Message : MinimumSigs.Message) : sig ... end

This module provides functions to display or interact with diagnostics in UNIX terminals.