asai: Compiler Diagnostics

Links

🔰 Quickstart tutorial

💡 Our design philosophy

📔 API reference

What is "asai"?

"asai" is the transliteration of "浅井", the family name of the character Kei Asai (浅井 ケイ) in the Japanese light novel Sagrada Reset (サクラダリセット, also known in English as Sakurada Reset). His ability is perfect photographic memory that is even immune to the "Reset" ability owned by another main character. This OCaml library should record all messages, just like the character.

Some Related Works

Rust Compiler Development Guide on Errors and Lints

Rust crate: miette

Rust crate: codespan

Rust crate: ariadne

OCaml development: grace